Add ReferenceCounter, MultiReferenceCounter
