+++ bigtable-hbase-mirroring-client-1.x-parent/bigtable-hbase-mirroring-client-1.x/src/main/java/com/google/cloud/bigtable/mirroring/hbase1_x/utils/referencecounting/MultiReferenceCounter.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.bigtable.mirroring.hbase1_x.utils.referencecounting;
+
+public class MultiReferenceCounter extends ReferenceCounter {
+  private final ReferenceCounter first;
+  private final ReferenceCounter second;
+
+  public MultiReferenceCounter(ReferenceCounter first, ReferenceCounter second) {
+    this.first = first;
+    this.second = second;
+  }
+
+  @Override
+  public void incrementReferenceCount() {
+    this.first.incrementReferenceCount();
+    this.second.incrementReferenceCount();
+  }
+
+  @Override
+  public void decrementReferenceCount() {
+    this.second.decrementReferenceCount();
+    this.first.decrementReferenceCount();
+  }
+}
